pyproject.toml
[tool.poetry]
name = "autobots"
version = "0.1.0"
description = ""
authors = ["Anurag Jha <toanuragjha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
uvicorn = "^0.23.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.0.3"}
loguru = "^0.7.0"
openai = {extras = ["datalib"], version = "^0.27.8"}
pytest-asyncio = "^0.21.1"
selenium = "4.11.2"
stability-sdk = "^0.8.3"
aiohttp = "^3.8.5"
pytest-xdist = "^3.3.1"
duckduckgo-search = "^3.8.4"
pinecone-client = {extras = ["grpc"], version = "^V3.0.0rc2"}
boto3 = "^1.28.0"
boto3-stubs = {extras = ["essential"], version = "^1.28.0"}
tiktoken = "^0.4.0"
pillow = "^10.0.0"
ddtrace = "^1.16.1"
pydantic-settings = "^2.0.2"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
supabase = "^1.0.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
redis = "^5.0.0"
pymongo = "^4.4.1"
aiodns = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
